Update LeetCode submission: Make Sum Divisible by P

+++ leetcode_submissions/2024-10-03/1694_Make_Sum_Divisible_by_P/Make_Sum_Divisible_by_P.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     int minSubarray(vector<int>& nums, int p) {
         int n = nums.size();
-        vector<long long> prefix(n+1, 0);  // Use long long for prefix sum to avoid overflow
-        long long total_sum = 0;  // Use long long for total sum
-        int ans = n;  // Start with the maximum possible subarray length, n
+        vector<int> prefix(n+1, 0); 
+        int total_sum = 0;
+        int ans = n;  
         
         // Step 1: Calculate prefix sum and total sum of the array
         for (int i = 0; i < n; i++) {
@@ -18,17 +18,18 @@ public:
         // Step 3: Brute force approach to check all subarrays
         for (int i = 0; i < n; i++) {
             for (int j = i; j < n; j++) {
-                
-                long long subarray_sum = prefix[j+1] - prefix[i];  // Sum of nums[i...j]
-                long long remaining_sum = total_sum - subarray_sum;
+                // Calculate the sum of the remaining elements if subarray nums[i...j] is removed
+                int subarray_sum = prefix[j+1] - prefix[i];  // Sum of nums[i...j]
+                int remaining_sum = total_sum - subarray_sum;
                 
                 // Step 4: Check if the remaining sum is divisible by p
                 if (remaining_sum % p == 0) {
-                    ans = min(ans, j - i + 1);  
+                    ans = min(ans, j - i + 1);  // Update the answer with the smallest subarray length
                 }
             }
         }
-      
+        
+        // Step 5: If we didn't find any valid subarray, return -1, otherwise return the answer
         return ans == n ? -1 : ans;
     }
 };
